Quick fixes to disable catalog functionality in Arduino build, so code will compile
